refactor: make AbstractXmlFilter
