--- src/main/kotlin/io/github/tacascer/namespace/NamespaceStrategy.kt
@@ -1,7 +1,0 @@
-package io.github.tacascer.namespace
-
-import org.jdom2.Element
-
-fun interface NamespaceStrategy {
-    fun process(element: Element)
-}
